feat: 자잘한 수정사항들 적용
- mongoDB 설정 수정
- 이미지 메타데이터도 DB에 저장
- 설정파일에 cloudFront 정보 추가

+++ src/main/kotlin/com/sheltersdog/image/entity/Image.kt
@@ -18,7 +18,7 @@ data class Image(
     val status: ImageStatus = ImageStatus.NOT_CONNECTED,
     val width: Int = 0,
     val height: Int = 0,
-    val size: Int = 0,
+    val size: Long = 0,
 
     val url: String = "",
     val regDate: LocalDateTime = LocalDateTime.now(),
